@@ -52,73 +56,149 @@ target-version = [
   "py312"
 ]
 
+[tool.bumpversion]
+commit = true
+commit_args = "--no-verify"
+tag = false
+allow_dirty = false
+parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)(\\-(?P<release>[a-z]+)(\\.(?P<build>\\d+)))?"
+serialize = [
+  "{major}.{minor}.{patch}-{release}.{build}",
+  "{major}.{minor}.{patch}"
+]
+
+[[tool.bumpversion.files]]
+filename = "glamod_marine_processing/__init__.py"
+search = "__version__ = \"{current_version}\""
+replace = "__version__ = \"{new_version}\""
+
+[tool.bumpversion.parts.build]
+independent = false
+
+[tool.bumpversion.parts.release]
+optional_value = "release"
+values = [
+  "dev",
+  "release"
+]
+
+[tool.codespell]
+ignore-words-list = "freez, fO, ND, BUID, SME, ASII, inpt"
+skip = "glamod_marine_processing/qc_suite/imma/*.json, tests/*.py, glamod_marine_processing/qc_suite/imma/code_tables/*"
+
 [tool.coverage.run]
 relative_files = true
 omit = ["tests/*.py"]
 
+[tool.deptry]
+extend_exclude = ["docs"]
+ignore_notebooks = true
+pep621_dev_dependency_groups = ["all", "dev", "docs"]
+
+[tool.deptry.package_module_name_map]
+"scikit-learn" = "sklearn"
+"pyyaml" = "yaml"
+
+[tool.deptry.per_rule_ignores]
+DEP001 = ["SBCK"]
+DEP002 = ["bottleneck", "pyarrow"]
+DEP004 = ["matplotlib", "pytest_socket"]
+
+[tool.flit.sdist]
+include = [
+  "AUTHORS.rst",
+  "CHANGES.rst",
+  "CONTRIBUTING.rst",
+  "LICENSE",
+  "Makefile",
+  "README.rst",
+  "ci/requirements/requirements.txt",
+  "ci/requirements/requirements_ob.txt",
+  "ci/requirements/requirements_qc.txt",
+  "docs/Makefile",
+  "docs/make.bat",
+  "tests/**/*.py"
+]
+exclude = [
+  "**/*.py[co]",
+  "**/__pycache__",
+  ".*",
+  "docs/_*",
+  "docs/modules.rst",
+  "pylintrc"
+]
+
 [tool.isort]
 profile = "black"
-py_version = 39
+py_version = 310
 append_only = true
 add_imports = "from __future__ import annotations"
 
 [tool.mypy]
-python_version = 3.9
+python_version = 3.10
 show_error_codes = true
 warn_return_any = true
-warn_unused_ignores = true
+warn_unused_configs = true
+plugins = ["numpy.typing.mypy_plugin"]
 
 [[tool.mypy.overrides]]
 module = [
-  "pandas",
-  "matplotlib",
-  "pytest"
+  "boltons.*",
+  "cftime.*",
+  "jsonpickle.*",
+  "numba.*",
+  "pytest_socket.*",
+  "SBCK.*",
+  "scipy.*",
+  "sklearn.*",
+  "statsmodels.*",
+  "yamale.*",
+  "yaml.*"
 ]
 ignore_missing_imports = true
-
-[tool.nbqa.md]
-mdformat = true
 
 [tool.pytest]
 python_files = "test_*.py"
 testpaths = ["./tests"]
 
-[tool.rstcheck]
-report_level = "WARNING"
-ignore_roles = [
-  "pull",
-  "issue"
-]
-
 [tool.ruff]
-src = ["cdm_reader_mapper"]
+src = ["glamod_marine_processing"]
 line-length = 150
-target-version = "py38"
-builtins = ["ellipsis"]
+target-version = "py39"
 exclude = [
   ".git",
-  ".eggs",
-  "docs",
   "build",
-  "tests"
+  ".eggs"
 ]
-extend-include = ["*.ipynb"]
+extend-include = [
+  "*.ipynb" # Include notebooks
+]
 
 [tool.ruff.format]
 line-ending = "auto"
 
 [tool.ruff.lint]
-ignore = [
-  "D205",
-  "D400",
-  "D401"
+extend-select = [
+  "RUF022" # unsorted-dunder-all
 ]
+ignore = [
+  "D205", # blank-line-after-summary
+  "D400", # ends-in-period
+  "D401", # non-imperative-mood
+  "S605", # starting process with shell
+  "S404", # subprocess module is insecure
+  "S602", # subprocess call with shell is True
+  "S603" # subprocess call with untrusted input
+]
+preview = true
 select = [
-  "C9",
-  "D",
-  "E",
-  "F",
-  "W"
+  "C90", # mccabe-complexity
+  "D", # docstrings
+  "E", # pycodestyle errors
+  "F", # pyflakes
+  "N802", # invalid-function-name
+  "S", # bandit
+  "W" # pycodestyle warnings
 ]
 
 [tool.ruff.lint.flake8-bandit]
