@@ -244,7 +244,7 @@ class imma:
 
     def md5(self, fname):
         """Return MD5 hash."""
-        hash_md5 = hashlib.md5()
+        hash_md5 = hashlib.md5()  # noqa: S324
         with open(fname, "rb") as f:
             for chunk in iter(lambda: f.read(4096), b""):
                 hash_md5.update(chunk)
@@ -314,7 +314,7 @@ class imma:
                 | (self.colTypes[column] == "float")
             ):
                 data[column] = self.numeric(data[column])
-                data = self.check_numeric_bad_data(data)
+                data = self.check_numeric_bad_data(data, column)
 
             if self.codetable[column] is not None:
                 # load code table
@@ -323,7 +323,7 @@ class imma:
                 ):
                     data[column] = self.to_str(data[column])
 
-                data = self.check_string_bad_data(data)
+                data = self.check_string_bad_data(data, column)
         return data
 
     def read_bufer(self, sections, block_size=None):
@@ -367,7 +367,7 @@ class imma:
             converters=self.conv,
         )
 
-    def loadImma(
+    def load_imma(
         self,
         filename,
         sections=["core", " 1", "98"],
