@@ -50,7 +50,8 @@ def modesp(awork):
         ifreq.append(0.0)
 
     ntime = len(awork)
-
+    atmode = 0
+    icmode = 0
     if ntime > 1:
         for i in range(1, ntime):
             # fixed so that indexing starts at zero
@@ -223,7 +224,7 @@ def distr2(invoyage):
 
     This takes the speed and direction reported by the ship and projects it forwards half a time step, it then projects
     it forwards another half time step using the speed and direction for the next report, to which the projected
-    location is then compared. The distances between the projeted and actual locations is returned
+    location is then compared. The distances between the projected and actual locations is returned
     """
     # Compute difference between actual and expected positions after
     # two observations IN REVERSE ORDER
@@ -516,6 +517,8 @@ def mds_track_check(invoyage):
     modal_speed = modesp(invoyage.get_speed())
     # set speed limits based on modal speed
     amax, amaxx, amin = set_speed_limits(modal_speed)
+    del amaxx
+    del amin
 
     # compare reported speeds and positions if we have them
     forward_diff_from_estimated = distr1(invoyage)
@@ -609,12 +612,12 @@ def mds_full_track_check(invoyage):
     """
     master_qc = mds_track_check(invoyage)
 
-    repititions = 0
+    repetitions = 0
 
     qcs = master_qc
 
     if len(qcs) > 0:
-        while max(qcs) > 0 and repititions < 4:
+        while max(qcs) > 0 and repetitions < 4:
             tempreps = ex.Voyage()
             qc_refs = []
 
@@ -630,7 +633,7 @@ def mds_full_track_check(invoyage):
             for i, qc_ref in enumerate(qc_refs):
                 master_qc[qc_ref] = qcs[i]
 
-            repititions += 1
+            repetitions += 1
 
     i = 0
     for rep in invoyage.rep_feed():
