@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pytest  # noqa
 
-import glamod_marine_processing
+import glamod_marine_processing  # noqa
 
 
 def test_marine_processing():
